Small changes to menu
 - Changed some values to make the player's movement cleaner
 - Renamed some things
 - Updated the save buttons
 - Added the create save screen
