+++ Reign of Yharim/Assets/Textures/GUI/TEX_CalamityTitle.png.meta
