+++ Reign of Yharim/Assets/Textures/GUI/TEX_Inventory.png.meta
