+++ Reign of Yharim/Assets/Textures/GUI/TEX_MenuBackground.png.meta
