+++ Reign of Yharim/Assets/Textures/GUI/TEX_ModeIndicatorDeath.png.meta
