+++ Reign of Yharim/Assets/Textures/GUI/TEX_ModeIndicatorInfernum.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7c542d6264c3fa84aa811f7784b0c65c
+guid: aa9c06cc434e5e847a5050c0b5ea52c5
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
@@ -33,7 +33,7 @@ TextureImporter:
   maxTextureSize: 2048
   textureSettings:
     serializedVersion: 2
-    filterMode: 1
+    filterMode: 0
     aniso: 1
     mipBias: 0
     wrapU: 1
@@ -47,7 +47,7 @@ TextureImporter:
   spriteMeshType: 1
   alignment: 0
   spritePivot: {x: 0.5, y: 0.5}
-  spritePixelsToUnits: 100
+  spritePixelsToUnits: 12
   spriteBorder: {x: 0, y: 0, z: 0, w: 0}
   spriteGenerateFallbackPhysicsShape: 1
   alphaUsage: 1
