+++ Reign of Yharim/Assets/Textures/GUI/TEX_ModeIndicatorNone.png.meta
