+++ Reign of Yharim/Assets/Textures/GUI/TEX_ModeIndicatorRev.png
