+++ Reign of Yharim/Assets/Textures/GUI/TEX_RoYTitle.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7c542d6264c3fa84aa811f7784b0c65c
+guid: 39932c195df2cad428055e0935511ef1
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
