+++ Reign of Yharim/UserSettings/EditorUserSettings.asset
@@ -6,10 +6,10 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
+      value: 51000c5e57570a585f5c5e214074074442161d282e7d71697b7e4a67b7b1616a
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 51000c5e57570a585f5c5e214074074442161d282e7d71697b7e4a67b7b1616a
+      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
